Prise en compte de l'ordre des axes selon le CRS pour les BBOX raster. Déplacement des fonctions publiques statiques de Raster vers Utils. Diverses corrections de style.
